restructure repo to new format, updated Cmake files
