+++ examples/Basic/test1
@@ -1,0 +1,1 @@
+#partitions = 10 bufsize = 10 count = 1 sum = 55.000000
