+++ src/bindings/CMakeLists.txt
@@ -1,4 +1,19 @@
 target_sources(MPIPCL
 	PRIVATE
-		"${CMAKE_CURRENT_SOURCE_DIR}/mpipcl.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Pready_range.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Request_free.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Test.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Testsome.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Waitany.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Parrived.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Precv_init.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Start.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Testall.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Wait.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Waitall.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Waitsome.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Pready.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Psend_init.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Startall.c"
+		"${CMAKE_CURRENT_SOURCE_DIR}/MPIP_Testany.c"
 )
